backend/requirements.txt
# Core backend frameworks
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Environment and utility
python-dotenv==1.1.0
gTTS==2.5.4
aiofiles==24.1.0

# MongoDB drivers
pymongo==4.13.0
motor==3.7.1

# Embeddings and vector search
sentence-transformers==4.1.0
faiss-cpu==1.8.0
numpy==1.26.4

# OpenAI API
openai==1.30.1

# Deepgram (speech-to-text)
deepgram-sdk==4.2.0
